requirements.txt
#client
python-socketio[client]
requests


#server
python-socketio
flask
cryptography
eventlet
flask-socketio

